Add category selection
Add 'collection' filed to the Feeds collection.
now user can create feed to any of the collection that is available. (previusely it could get only Prosucts)
the collections list is not dynamic and is written specifc in the code. it seems that you cant get the collection's list from the payload.config dynamicly